* SPR-9431:
  Update -Xlint compiler warning output
  Fix various compiler warnings in spring-context
  Fix "unnecessary @SuppressWarnings" warnings
  Fix [rawtypes] compiler warnings
  Fix [dep-ann] warnings with @Deprecated
  Fix [cast] compiler warnings
  Fix [serial] compiler warnings
  Fix [varargs] compiler warnings
  Fix warnings due to unused import statements

+++ spring-context/src/test/java/org/springframework/jmx/export/assembler/AbstractJmxAssemblerTests.java
@@ -145,14 +145,14 @@ public abstract class AbstractJmxAssemblerTests extends AbstractJmxTests {
 		ModelMBeanOperationInfo get = info.getOperation("getName");
 		assertNotNull("get operation should not be null", get);
 		assertEquals("get operation should have visibility of four",
-				(Integer) get.getDescriptor().getFieldValue("visibility"),
+				get.getDescriptor().getFieldValue("visibility"),
 				new Integer(4));
 		assertEquals("get operation should have role \"getter\"", "getter", get.getDescriptor().getFieldValue("role"));
 
 		ModelMBeanOperationInfo set = info.getOperation("setName");
 		assertNotNull("set operation should not be null", set);
 		assertEquals("set operation should have visibility of four",
-				(Integer) set.getDescriptor().getFieldValue("visibility"),
+				set.getDescriptor().getFieldValue("visibility"),
 				new Integer(4));
 		assertEquals("set operation should have role \"setter\"", "setter", set.getDescriptor().getFieldValue("role"));
 	}
